+++ .github/scripts/main-publish.sh
@@ -1,0 +1,4 @@
+# publish packages
+pnpm ci:publish
+# enter prerelease mode until the next main release
+pnpm changeset pre enter canary