+++ packages/nextjs-kit/src/components/index.ts
@@ -1,3 +1,6 @@
 export { default as Header } from './header';
 export { default as Footer } from './footer';
 export { default as Paginator } from './paginator';
+export { default as ContentWithImage } from './contentWithImage';
+export { default as Recipe } from './recipe';
+export { Grid, withGrid } from './grid';