+++ packages/wordpress-kit/src/lib/index.ts
@@ -1,0 +1,3 @@
+export { default as tailwindcssPlugin } from './tailwindcssPlugin';
+export { default as GraphqlClientFactory } from './GraphqlClientFactory';
+export { default as setEdgeHeader } from './setEdgeHeader';