+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/components/index.ts
@@ -5,6 +5,8 @@ import { Quote } from './Quote';
 import { GalleryComponent } from './Gallery';
 import { AudioComponent } from './Audio';
 import { CoverComponent } from './Cover';
+import { mediaAndText } from './MediaAndText';
+import { ButtonsComponent } from './Buttons';
 
 export {
 	ImageComponent,
@@ -14,4 +16,6 @@ export {
 	GalleryComponent,
 	AudioComponent,
 	CoverComponent,
+	mediaAndText,
+	ButtonsComponent,
 };